+++ validator/BUILD.bazel
@@ -21,7 +21,7 @@ go_library(
         "//shared/logutil:go_default_library",
         "//shared/params:go_default_library",
         "//shared/version:go_default_library",
-        "//validator/accounts:go_default_library",
+        "//validator/accounts/v1:go_default_library",
         "//validator/client/streaming:go_default_library",
         "//validator/flags:go_default_library",
         "//validator/node:go_default_library",
@@ -62,7 +62,7 @@ go_image(
         "//shared/logutil:go_default_library",
         "//shared/params:go_default_library",
         "//shared/version:go_default_library",
-        "//validator/accounts:go_default_library",
+        "//validator/accounts/v1:go_default_library",
         "//validator/client/polling:go_default_library",
         "//validator/client/streaming:go_default_library",
         "//validator/flags:go_default_library",
