+++ validator/accounts/v1/account.go
@@ -1,5 +1,5 @@
 // Package accounts defines tools to manage an encrypted validator keystore.
-package accounts
+package v1
 
 import (
 	"bufio"
