+++ validator/accounts/v1/status_test.go
@@ -1,4 +1,4 @@
-package accounts
+package v1
 
 import (
 	"context"
@@ -6,6 +6,7 @@ import (
 
 	"github.com/golang/mock/gomock"
 	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
+
 	"github.com/prysmaticlabs/prysm/shared/mock"
 )
