+++ validator/client/polling/BUILD.bazel
@@ -81,7 +81,7 @@ go_test(
         "//shared/roughtime:go_default_library",
         "//shared/slotutil:go_default_library",
         "//shared/testutil:go_default_library",
-        "//validator/accounts:go_default_library",
+        "//validator/accounts/v1:go_default_library",
         "//validator/db:go_default_library",
         "//validator/keymanager:go_default_library",
         "@com_github_gogo_protobuf//types:go_default_library",
