+++ validator/keymanager/BUILD.bazel
@@ -21,7 +21,7 @@ go_library(
         "//shared/bytesutil:go_default_library",
         "//shared/interop:go_default_library",
         "//shared/params:go_default_library",
-        "//validator/accounts:go_default_library",
+        "//validator/accounts/v1:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
