+++ validator/node/BUILD.bazel
@@ -8,7 +8,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//shared/testutil:go_default_library",
-        "//validator/accounts:go_default_library",
+        "//validator/accounts/v1:go_default_library",
         "@com_github_urfave_cli_v2//:go_default_library",
     ],
 )
